BREAKING CHANGE: 💣 Reorganize exceptions Signed-off-by: Samuel Wu <twopizza9621536@gmail.com>
