--- src/legacy_puyo_tools/exceptions.py
@@ -1,10 +1,0 @@
-# SPDX-FileCopyrightText: 2021 Nick Woronekin
-# SPDX-FileCopyrightText: 2025 Samuel Wu
-#
-# SPDX-License-Identifier: MIT
-
-"""Exceptions that might be thrown when parsing files for converters."""
-
-
-class FormatError(Exception):
-    """The current input data does not conform to an expected format."""
